Move symbols to base directory
