+++ src/qibo/hamiltonians/__init__.py
@@ -1,3 +1,2 @@
 from qibo.hamiltonians.hamiltonians import *
-from qibo.hamiltonians.symbols import *
 from qibo.hamiltonians.models import X, Y, Z, XXZ, TFIM, MaxCut
